requirements.txt
streamlit>=1.36.0
pdfplumber>=0.11.0
PyMuPDF>=1.24.1
pdfminer.six>=20231228
openai>=1.40.0
anthropic>=0.34.2
python-dotenv>=1.0.1
plotly>=5.23.0
pandas>=2.2.2


